Add files via upload
There are two computers that act as client and server. The client and server communicate using TCP socket. The client will send a message to the server. The message will be sent back to the client by the server. The client will receive a reply message from the server and display it.